% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-low-freq.R
\name{check_low_freq}
\alias{check_low_freq}
\alias{check_low_freq2}
\title{Check for low frequencies/counts when doing logistic regression}
\usage{
check_low_freq(fit, data, threshold = 0.01)

check_low_freq2(data, outcome, predictors, threshold = 0.01)
}
\arguments{
\item{fit}{An object of class \code{glm} inheriting from "glm" which inherits from
the class "lm.}

\item{data}{A tibble or data frame with the full data set.}

\item{threshold}{The threshold to flag categories with frequencies/counts.
Default is 0.01.}

\item{outcome}{Character string. The dependent variable (outcome) for
logistic regression.}

\item{predictors}{Character vector. Independent variables
(predictors/covariates) for univariable and/or multivariable modelling.}
}
\value{
A tibble
}
\description{
Low counts in certain categories and levels of a variable can cause issues in
logistic regression. This function helps to identify low counts that might be
problematic. Rule of thumb might be to collapse categories that contain <=
1.0\% of your data.
}
\examples{
\dontrun{
library(epiDisplay)
library(dplyr)

dplyr::glimpse(infert)
model0 <- glm(case ~ induced + spontaneous + education,
              family = binomial,
              data = infert)
summary(model0)


check_low_freq(fit = model0,
               data = infert)

check_low_freq(fit = model0,
               data = infert, 
               threshold = 0.05)

check_low_freq2(data = infert,
                outcome = "case",
                predictors = c("induced", "spontaneous", "education"), 
                threshold = 0.05)


#### Another data set --------------------------------

library(compareGroups)
data(predimed)
dplyr::glimpse(predimed)
predimed <- predimed \%>\%
  mutate_if(is.double, as.double)

fit = glm(htn ~ sex + bmi + smoke,
          family = binomial(link = "logit"),
          data = predimed)

check_low_freq(fit = fit,
               data = predimed)

check_low_freq(fit = model0,
               data = infert, 
               threshold = 0.05)

check_low_freq2(data = predimed,
                outcome = "htn",
                predictors = c("sex", "bmi", "smoke"), 
                threshold = 0.01)


check_low_freq2(data = predimed,
                outcome = "htn",
                predictors = c("sex", "bmi", "smoke"), 
                threshold = 0.05)



}
}
