% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-logistic.R
\name{display_logistic}
\alias{display_logistic}
\title{Display Logistic Regression Results}
\usage{
display_logistic(
  fit,
  data,
  add_multi = FALSE,
  format = "html",
  conf_level = 0.95,
  exponentiate = TRUE,
  include_last_row = TRUE
)
}
\arguments{
\item{fit}{An object of class \code{glm} with \code{family = binomial}.}

\item{data}{A tibble or data frame with the full dataset used for modeling.}

\item{add_multi}{Logical; whether to include multivariable results. Default
is \code{FALSE}.}

\item{format}{Display format. Default is \code{"html"}.}

\item{conf_level}{Confidence level for intervals. Default is \code{0.95}.}

\item{exponentiate}{Logical; exponentiate estimates to ORs. Default is \code{TRUE}.}

\item{include_last_row}{Logical; add blank spacing row. Default is \code{TRUE}.}
}
\value{
A tibble containing the logistic regression results.
}
\description{
Given a \code{glm} model fit object from logistic regression, displays
univariable and (optionally) multivariable results in a format suitable for
presentation.
}
\examples{
\dontrun{
dplyr::glimpse(infert)

fit <- glm(case ~ induced + spontaneous + education,
           family = binomial,
           data = infert)

display_logistic(fit, data = infert)
display_logistic(fit, data = infert, add_multi = TRUE)

# Interaction
fit_int <- glm(case ~ induced * education,
               family = binomial,
               data = infert)
display_logistic(fit_int, data = infert, add_multi = TRUE)
}

}
