% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latable.R
\docType{package}
\name{latable-package}
\alias{latable}
\alias{latable-package}
\title{latable: Tools for making tables of results}
\description{
This package contains a variety of helper functions for making tables of
results easily.
}
\examples{
# Example usage:
library(latable)

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emilelatour/latable}
  \item Report bugs at \url{https://github.com/emilelatour/latable/issues}
}

}
\author{
\strong{Maintainer}: Emile Latour \email{latour@ohsu.edu}

}
\keyword{internal}
