% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-coxph.R
\name{display_coxph2}
\alias{display_coxph2}
\title{Display Cox Proportional Hazards Regression Results (given variable names)}
\usage{
display_coxph2(
  data,
  outcome,
  predictors,
  add_multi = FALSE,
  format = "html",
  conf_level = 0.95,
  exponentiate = TRUE,
  include_last_row = TRUE
)
}
\arguments{
\item{data}{A tibble or data frame with the full dataset.}

\item{outcome}{Character string. The dependent variable (outcome) for
Cox regression, formatted as "Surv(time, event)".}

\item{predictors}{Character vector. Independent variables
(predictors/covariates) for univariable and/or multivariable modeling.}

\item{add_multi}{Logical; include multivariable results. Default is \code{FALSE}}

\item{format}{Display format in case characters need escaping. Placeholder for future use. Default is "html".}

\item{conf_level}{The confidence level to use for confidence intervals.
Must be between 0 and 1. Default is 0.95.}

\item{exponentiate}{Logical; whether to exponentiate coefficient estimates to show hazard ratios. Default is \code{TRUE}.}

\item{include_last_row}{Adds a row at the end of each set of results for spacing. Default is \code{TRUE}.}
}
\value{
A tibble containing Cox regression results.
}
\description{
Given a string with an "outcome" and a vector string with "predictors",
displays univariable and (optionally) multivariable results from Cox proportional hazards regression
in a format suitable for presentation.

This version allows you to explicitly provide the outcome and predictors in
case there are issues with getting a full model fit.
}
\examples{
\dontrun{
library(survival)
data(lung)

display_coxph2(data = lung,
               outcome = "Surv(time, status)",
               predictors = c("age", "sex", "ph.ecog"))

display_coxph2(data = lung,
               outcome = "Surv(time, status)",
               predictors = c("age", "sex", "ph.ecog"),
               add_multi = TRUE)
}
}
