% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-linear.R
\name{display_linear}
\alias{display_linear}
\title{Display linear regression results (given a model fit)}
\usage{
display_linear(
  fit,
  data,
  add_multi = FALSE,
  format = "html",
  conf_level = 0.95,
  include_last_row = TRUE
)
}
\arguments{
\item{fit}{An object of class \code{lm} or \code{glm} with \code{family = gaussian}.}

\item{data}{A tibble or data frame with the full data set.}

\item{add_multi}{Logical; include multivariable results. Default is \code{FALSE}.}

\item{format}{Display format in case I need to escape some characters. A
placeholder for now in case I need it in the future. Default is "html".}

\item{conf_level}{The confidence level to use for the confidence interval.
Must be strictly greater than 0 and less than 1. Defaults to 0.95, which
corresponds to a 95 percent confidence interval.}

\item{include_last_row}{Adds a row at the end of each set of results to give
some breathing room. Default is \code{TRUE}.}
}
\value{
A tibble
}
\description{
Given a \code{fit} object, univariable and (optionally) multivariable results from
linear regression are displayed in a format suitable for presentation.
}
\examples{
\dontrun{
# Example with the `mtcars` dataset:
model <- glm(mpg ~ cyl + disp, family = gaussian, data = mtcars)
display_linear(fit = model, data = mtcars, add_multi = TRUE)
}
}
