% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-linear.R
\name{display_linear2}
\alias{display_linear2}
\title{Display linear regression results (given variable names)}
\usage{
display_linear2(
  data,
  outcome,
  predictors,
  add_multi = FALSE,
  format = "html",
  conf_level = 0.95,
  include_last_row = TRUE
)
}
\arguments{
\item{data}{A tibble or data frame with the full data set.}

\item{outcome}{Character string. The dependent variable (outcome) for
linear regression.}

\item{predictors}{Character vector. Independent variables
(predictors/covariates) for univariable and/or multivariable modelling.}

\item{add_multi}{Logical; include multivariable results. Default is \code{FALSE}.}

\item{format}{Display format in case I need to escape some characters. A
placeholder for now in case I need it in the future. Default is "html".}

\item{conf_level}{The confidence level to use for the confidence interval.
Must be strictly greater than 0 and less than 1. Defaults to 0.95, which
corresponds to a 95 percent confidence interval.}

\item{include_last_row}{Adds a row at the end of each set of results to give
some breathing room. Default is \code{TRUE}.}
}
\value{
A tibble with regression results.
}
\description{
Given a string with an "outcome" and a vector string with "predictors",
univariable and (optionally) multivariable results from linear regression
are displayed in a format suitable for presentation.

This version allows you to explicitly provide the outcome and predictors in
case there are issues with getting a full model fit.
}
\examples{
\dontrun{
display_linear2(data = mtcars,
                outcome = "mpg",
                predictors = c("cyl", "disp"))

display_linear2(data = mtcars,
                outcome = "mpg",
                predictors = c("cyl", "disp"),
                add_multi = TRUE)
}
}
