% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-logistic2.R
\name{display_logistic2}
\alias{display_logistic2}
\title{Display logistic regression results (given variable names)}
\usage{
display_logistic2(
  data,
  outcome,
  predictors,
  ref_level = NULL,
  add_multi = FALSE,
  format = "html",
  conf_level = 0.95,
  exponentiate = TRUE,
  include_last_row = TRUE
)
}
\arguments{
\item{data}{A tibble or data frame with the full data set.}

\item{outcome}{Character string. The dependent variable (outcome) for
logistic regression.}

\item{predictors}{Character vector. Independent variables
(predictors/covariates) for univariable and/or multivariable modelling.}

\item{ref_level}{Character string. The factor level of outcome variable that
corresponds to the true condition (1). If not provided then default is
\code{NULL} and the model fit will determine the referenc level.}

\item{add_multi}{Logical; include multivariable results. Default is \code{FALSE}}

\item{format}{Display format in case I need to escape some characters. A
place holder for now in case I need it in the future. Default is "html".}

\item{conf_level}{The confidence level to use for the confidence interval.
Must be strictly greater than 0 and less than 1. Defaults to 0.95, which
corresponds to a 95 percent confidence interval.}

\item{exponentiate}{Logical indicating whether or not to exponentiate the
the coefficient estimates. This is typical for logistic and multinomial
regressions, but a bad idea if there is no log or logit link. Defaults to
\code{TRUE}.}

\item{include_last_row}{Adds a row at the end of each set of results to give
some breathing room. Default is \code{TRUE}.}
}
\value{
A tibble
}
\description{
Given a string with an "outcome" and a vector string with "predictors",
univariable and (optionally) multivariable results from logistic regression
are displayed in a format suitable for presentation.

This version allows you to explicitly provide the outcome and predictors in
case there are issues with getting a full model fit.
}
\examples{
\dontrun{
library(epiDisplay)
library(tidyverse)
dplyr::glimpse(infert)

case ~ induced + spontaneous + education

display_logistic2(data = infert,
                  outcome = "case",
                  predictors = c("induced", "spontaneous", "education"))

display_logistic2(data = infert,
                  outcome = "case",
                  predictors = c("induced", "spontaneous", "education"),
                  add_multi = TRUE)

# How does an interaction term work?
# Doesn't work yet (2019-03-23)
# display_logistic2(data = infert,
#                   outcome = "case",
#                   predictors = c("induced * education"))

#### Another data set --------------------------------
library(compareGroups)
data(predimed)
dplyr::glimpse(predimed)
predimed <- predimed \%>\%
  mutate_if(is.double, as.double)

fit = glm(htn ~ sex + bmi + smoke,
          family = binomial(link = "logit"),
          data = predimed)

display_logistic2(data = predimed,
                  outcome = "htn",
                  predictors = c("sex", "bmi", "smoke"),
                  add_multi = TRUE)
 }
}
