% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-linear.R
\name{.do_linear_univ}
\alias{.do_linear_univ}
\title{Do univariable linear regression and extract results in nice format.}
\usage{
.do_linear_univ(
  data,
  formula,
  format = "html",
  conf_level = 0.95,
  include_last_row = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{formula}{A character string specifying the formula for the regression.}

\item{format}{Display format in case I need to escape some characters. A
placeholder for now in case I need it in the future. Default is "html".}

\item{conf_level}{The confidence level to use for the confidence interval.
Must be strictly greater than 0 and less than 1. Defaults to 0.95, which
corresponds to a 95 percent confidence interval.}

\item{include_last_row}{Adds a row at the end of each set of results to give
some breathing room. Default is \code{TRUE}.}

\item{...}{Additional arguments.}
}
\value{
A tibble or data frame with regression results.
}
\description{
A helper function to be used in a loop to perform univariable linear regression
and output nicely formatted results.
}
