% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latable.R
\docType{package}
\name{latable}
\alias{latable}
\alias{latable-package}
\title{latable: Tools for making tables of results}
\description{
This package contains a variety of helper functions for making tables of
results easily.
}
\examples{
# Example usage:
library(latable)

}
